+++ pugixml/utils.cpp
@@ -32,7 +32,7 @@ void pugixml_utils_error_logger (void * pugi_xml_parse_result, const string & xm
   if (result->status == status_ok) return;
   int start = static_cast<int>(result->offset - 10);
   if (start < 0) start = 0;
-  string fragment = xml.substr (start, 20);
+  string fragment = xml.substr (static_cast<size_t>(start), 20);
   fragment = filter_string_str_replace ("\n", "", fragment);
   string msg;
   msg.append (result->description());
