+++ sources/etcbc4.cpp
@@ -77,9 +77,9 @@ void sources_etcbc4_download ()
     "Maleachi"
   };
 
-  for (unsigned int bk = 0; bk < books.size (); bk++) {
+  for (size_t bk = 0; bk < books.size (); bk++) {
 
-    int book = bk + 1;
+    int book = static_cast<int>(bk + 1);
     string bookname = books[bk];
 
     bool book_done = false;
